Extract Text.GetStyledSubstrings into a shared StyledSubstringSplitter class
Pulls the BBCode inline-variable run-splitting algorithm out of the MonoGame-only Text renderable into RenderingLibrary/Graphics/StyledSubstringSplitter.cs. It has no MonoGame dependency (System.Drawing.Color, plain lists), so other backends can reuse it instead of reimplementing run-splitting - prep for raylib BBCode support (#3471).

Behavior-preserving: Text.GetStyledSubstrings delegates to the new class, and the existing pinning tests (TextRuntimeTests) pass unchanged.
