+++ MonoGameGum/MonoGameGum.csproj
@@ -227,6 +227,7 @@
 		<Compile Include="..\RenderingLibrary\Graphics\SpriteBatchStack.cs" Link="RenderingLibrary\SpriteBatchStack.cs" />
 		<Compile Include="..\RenderingLibrary\Graphics\SpriteManager.cs" Link="RenderingLibrary\SpriteManager.cs" />
 		<Compile Include="..\RenderingLibrary\Graphics\SpriteRenderer.cs" Link="RenderingLibrary\SpriteRenderer.cs" />
+		<Compile Include="..\RenderingLibrary\Graphics\StyledSubstringSplitter.cs" Link="Renderables\StyledSubstringSplitter.cs" />
 		<Compile Include="..\RenderingLibrary\Graphics\Text.cs" Link="Renderables\Text.cs" />
 		<Compile Include="..\RenderingLibrary\Graphics\TextExtensions.cs" Link="Renderables\TextExtensions.cs" />
 		<Compile Include="..\RenderingLibrary\Graphics\TextManager.cs" Link="RenderingLibrary\TextManager.cs" />
